+++ src/components/Forms/SignUp/SignUp.js
@@ -1,8 +1,8 @@
 import React, { useRef, useState } from 'react'
 import { Card, Form, Button, Alert } from 'react-bootstrap'
 import { useAuth } from '../../../context/AuthContext'
-import { Link, useNavigate } from 'react-router-dom'
-
+import Link from 'next/link'
+import { useRouter } from 'next/router'
 import { ToastContainer, toast } from 'react-toastify'
 
 export default function SignUp() {
@@ -12,7 +12,7 @@ export default function SignUp() {
   const confirmPasswordRef = useRef()
   /*   const [error, setError] = useState('') */
   const [loading, setLoading] = useState(false)
-  const navigate = useNavigate()
+  const router = useRouter()
 
   const { signup } = useAuth()
 
@@ -34,7 +34,7 @@ export default function SignUp() {
       )
       /* console.log('account created!') */
       toast('account created!')
-      navigate('/kiosk')
+      router.push('/kiosk')
     } catch {
       /* setError('Error while creating an account') */
       toast('Error while creating an account')
@@ -77,7 +77,7 @@ export default function SignUp() {
           </Form>
           <div className='m-2'>
             Already have an account?{' '}
-            <Link className='myLink' to='/'>
+            <Link className='myLink' href='/'>
               Let's go.
             </Link>
           </div>