+++ src/components/kiosk/nyt/MovieReviews.js
@@ -6,7 +6,7 @@ import PostCardMovies from '../../PostCardMovies'
 import { useAuth } from '../../../context/AuthContext'
 
 import data from '../../../data/nytMovieReviews.json'
-const nytMovieReviewsUrl = `https://api.nytimes.com/svc/movies/v2/reviews/picks.json?&api-key=${process.env.REACT_APP_NYT_API_KEY}`
+const nytMovieReviewsUrl = `https://api.nytimes.com/svc/movies/v2/reviews/picks.json?&api-key=${process.env.NEXT_PUBLIC_NYT_API_KEY}`
 
 export default function MovieReviews() {
   const [reviews, setReviews] = useState([])