+++ src/components/kiosk/nyt/TopStories/ArtTopStories.js
@@ -5,7 +5,7 @@ import PostCard from '../../../PostCard'
 import { useAuth } from '../../../../context/AuthContext'
 
 import data from '../../../../data/nytArt.json'
-const nytTopStoriesUrl = `https://api.nytimes.com/svc/topstories/v2/arts.json?api-key=${process.env.REACT_APP_NYT_API_KEY}`
+const nytTopStoriesUrl = `https://api.nytimes.com/svc/topstories/v2/arts.json?api-key=${process.env.NEXT_PUBLIC_NYT_API_KEY}`
 /* allowed values: arts, automobiles, books, business, fashion, food, health, home, insider, magazine, movies, nyregion, obituaries, opinion, politics, realestate, science, sports, sundayreview, technology, theater, t-magazine, travel, upshot, us, world */
 
 export default function ArtTopStories() {