+++ src/components/kiosk/nyt/TopStories/PoliticsTopStories.js
@@ -6,7 +6,7 @@ import { useAuth } from '../../../../context/AuthContext'
 
 import data from '../../../../data/nytPolitics.json'
 
-const nytTopStoriesUrl = `https://api.nytimes.com/svc/topstories/v2/politics.json?api-key=${process.env.REACT_APP_NYT_API_KEY}`
+const nytTopStoriesUrl = `https://api.nytimes.com/svc/topstories/v2/politics.json?api-key=${process.env.NEXT_PUBLIC_NYT_API_KEY}`
 /* allowed values: arts, automobiles, books, business, fashion, food, health, home, insider, magazine, movies, nyregion, obituaries, opinion, politics, realestate, science, sports, sundayreview, technology, theater, t-magazine, travel, upshot, us, world */
 
 export default function PoliticsTopStories() {